examples/python/requirements.txt
grpcio==1.6.0
grpcio-tools==1.6.0
protobuf==3.4.0
sanction==0.4.1
six==1.11.0
